[Silver I] Title: 동물원, Time: 8 ms, Memory: 7532 KB -BaekjoonHub

+++ 백준/Silver/1309. 동물원/동물원.cc
@@ -10,13 +10,13 @@ int main(void){
     int N;
     cin>>N;//사자 우리의 크기
 
-    vector<vector<long> > DP(N+1,vector<long> (2));
+    vector<vector<int> > DP(N+1,vector<int> (2));
 
     DP[0][0]=DP[0][1]=0;
     DP[1][0]=1;
     DP[1][1]=2;
 
-    long result;
+    int result;
 
     for(int i=2; i<N+1; i++){
         DP[i][0]=(DP[i-1][0]+DP[i-1][1])%9901;
